+++ aria_data_loaders/aria_data_utils/detector_wrappers/generic_detector_interface.py
@@ -1,0 +1,15 @@
+import numpy as np
+
+
+class GenericDetector:
+    def __init__(self):
+        self.is_enabled = False
+
+    def enable_detector(self):
+        self.is_enabled = True
+
+    def disable_detector(self):
+        self.is_enabled = False
+
+    def process_frame(self, frame: np.ndarray):
+        raise NotImplementedError
